pruning the repo and remain `parse_file` only
