--- llm_config/config/templates/test_model.ini
@@ -1,28 +1,0 @@
-[Model]
-name=Test-Model
-num_layers=32
-emb_dim=4096
-max_seq_len=8192
-vocab_size=128256
-tie_word_embeddings=false
-
-[MHA]
-name=Test-Model
-out_projection_use_bias=false
-dropout=0.0
-num_query_heads=32
-num_kv_heads=8
-
-[Rope]
-name=Test-Model
-theta=500000.0
-type=rope
-
-[MLP]
-name=Test-Model
-activation=silu
-dim=14336
-
-[Norm]
-name=Test-Model
-epsilon=1e-05
