--- tests/__init__.py
@@ -1,2 +1,0 @@
-# Tests package
-
