--- tests/templates/test-llama-3-8B.ini
@@ -1,28 +1,0 @@
-[Model]
-name=Test-Llama-3-8B
-num_layers=32
-emb_dim=4096
-max_seq_len=8192
-vocab_size=128256
-tie_word_embeddings=false
-
-[MHA]
-name=Test-Llama-3-8B
-out_projection_use_bias=false
-dropout=0.0
-num_query_heads=32
-num_kv_heads=8
-
-[Rope]
-name=Test-Llama-3-8B
-theta=500000.0
-type=rope
-
-[MLP]
-name=Test-Llama-3-8B
-activation=silu
-dim=14336
-
-[Norm]
-name=Test-Llama-3-8B
-epsilon=1e-05
