--- tests/templates/test-mla.ini
@@ -1,37 +1,0 @@
-# for test
-[Custom1]
-name=Test-Custom1
-out_projection_use_bias=false
-dropout=0.0
-kv_lora_rank=512
-q_lora_rank=1536
-qk_nope_head_dim=128
-qk_rope_head_dim=64
-v_head_dim=128
-num_query_heads=128
-num_kv_heads=128
-
-[Custom2]
-name=Test-Custom2
-out_projection_use_bias=false
-dropout=0.0
-kv_lora_rank=512
-q_lora_rank=1536
-qk_nope_head_dim=128
-qk_rope_head_dim=64
-v_head_dim=128
-num_query_heads=128
-num_kv_heads=128
-
-[Custom3]
-name=Test-Custom3
-out_projection_use_bias=false
-dropout=0.0
-kv_lora_rank=512
-q_lora_rank=1536
-qk_nope_head_dim=128
-qk_rope_head_dim=64
-v_head_dim=128
-num_query_heads=128
-num_kv_heads=128
-
