--- tests/templates/test-moe.ini
@@ -1,15 +1,0 @@
-[MoE]
-name=Test-CUSTOM
-aux_loss_alpha=0.001
-n_routed_experts=256
-n_shared_experts=1
-n_activated_experts=8
-norm_topk_prob=true
-score_func=sigmoid
-route_scale=2.5
-seq_aux=true
-shared_experts_activation=silu
-shared_experts_dim=2048
-routed_experts_activation=silu
-routed_experts_dim=2048
-shared_experts_use_gate=true
