--- tests/templates/test_deepseek-v3-671B.ini
@@ -1,57 +1,0 @@
-[Model]
-name=Test-DeepSeek-V3-671B
-num_layers=61
-emb_dim=7168
-n_dense_layers=3
-max_seq_len=163840
-vocab_size=129280
-tie_word_embeddings=false
-
-[MLA]
-name=Test-DeepSeek-V3-671B
-out_projection_use_bias=false
-dropout=0.0
-kv_lora_rank=512
-q_lora_rank=1536
-qk_nope_head_dim=128
-qk_rope_head_dim=64
-v_head_dim=128
-num_query_heads=128
-num_kv_heads=128
-
-[Rope]
-name=Test-DeepSeek-V3-671B
-theta=10000
-beta_fast=32
-beta_slow=1
-factor=40
-original_seq_len=4096
-mscale=1.0
-mscale_all_dim=1.0
-type=yarn
-
-[MoE]
-name=Test-DeepSeek-V3-671B
-aux_loss_alpha=0.001
-n_routed_experts=256
-n_shared_experts=1
-n_activated_experts=8
-norm_topk_prob=true
-score_func=sigmoid
-route_scale=2.5
-seq_aux=true
-shared_experts_activation=silu
-shared_experts_dim=2048
-routed_experts_activation=silu
-routed_experts_dim=2048
-shared_experts_use_gate=true
-
-[MLP]
-name=Test-DeepSeek-V3-671B-Dense
-activation=silu
-dim=18432
-use_gate=true
-
-[Norm]
-name=Test-DeepSeek-V3-671B
-epsilon=1e-06
